+++ RawTools/Algorithms/Extraction/ExtractData.cs
@@ -687,6 +687,12 @@ namespace RawTools.Algorithms.ExtractData
                 sps2.CopyTo(spsMasses, sps1.Length);
             }
             trailerExtra.SPSMasses = spsMasses;
+
+            if (indices.FaimsVoltage != -1)
+            {
+                trailerExtra.FaimsVoltage = Convert.ToDouble(rawFile.GetTrailerExtraValue(scan, indices.FaimsVoltage));
+            }
+
             return trailerExtra;
         }
 
@@ -763,7 +769,7 @@ namespace RawTools.Algorithms.ExtractData
 
             methodData.CreationDate = rawFile.CreationDate;
             methodData.Instrument = rawFile.GetInstrumentData().Name;
-
+            
             methodData.AnalysisOrder = index.AnalysisOrder;
             int firstQuantScan = index.ScanEnumerators[index.AnalysisOrder][0];
             methodData.QuantAnalyzer = index.allScans[firstQuantScan].MassAnalyzer;
@@ -797,5 +803,29 @@ namespace RawTools.Algorithms.ExtractData
 
             return methodData;
         }
+
+        public static MethodDataContainer Ms1MethodData(IRawDataPlus rawFile, ScanIndex index)
+        {
+            rawFile.SelectInstrument(Device.MS, 1);
+
+            Log.Information("Extracting method/instrument information");
+
+            MethodDataContainer methodData = new MethodDataContainer();
+
+            methodData.CreationDate = rawFile.CreationDate;
+            methodData.Instrument = rawFile.GetInstrumentData().Name;
+
+            methodData.AnalysisOrder = index.AnalysisOrder;
+            int firstQuantScan = index.ScanEnumerators[index.AnalysisOrder][0];
+            methodData.QuantAnalyzer = index.allScans[firstQuantScan].MassAnalyzer;
+
+            int firstMs1Scan = index.ScanEnumerators[MSOrderType.Ms][0];
+            methodData.MassAnalyzers.Add(MSOrderType.Ms, index.allScans[firstMs1Scan].MassAnalyzer);
+
+            methodData.MSOrderEnumerator.Add(MSOrderType.Ms);
+
+            return methodData;
+        }
     }
+
 }