+++ RawTools/Properties/AssemblyInfo.cs
@@ -36,4 +36,4 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyFileVersion("15.1.2019.0")]
 
-[assembly: AssemblyInformationalVersion("2.0.4")]
+[assembly: AssemblyInformationalVersion("2.0.5")]