+++ RawToolsGUI/RawToolsGUI.cs
@@ -150,10 +150,12 @@ namespace RawToolsGUI
             if (checkBoxModeParse.Checked)
             {
                 groupBoxDataOutput.Enabled = true;
+                groupBoxCommonOptions.Enabled = true;
             }
             else
             {
                 groupBoxDataOutput.Enabled = false;
+                groupBoxCommonOptions.Enabled = false;
             }
         }
 
@@ -165,7 +167,7 @@ namespace RawToolsGUI
                 radioButtonSelectDirectory.Enabled = true;
                 buttonSelectDirectory.Enabled = true;
                 textBoxRawFileDirectory.Enabled = true;
-
+                groupBoxCommonOptions.Enabled = true;
                 radioButtonSelectFiles.Checked = false;
                 radioButtonSelectFiles.Enabled = false;
                 buttonSelectFiles.Enabled = false;
@@ -179,13 +181,29 @@ namespace RawToolsGUI
                 radioButtonSelectDirectory.Enabled = true;
                 buttonSelectDirectory.Enabled = true;
                 textBoxRawFileDirectory.Enabled = true;
-
+                groupBoxCommonOptions.Enabled = false;
                 radioButtonSelectFiles.Checked = false;
                 radioButtonSelectFiles.Enabled = true;
                 buttonSelectFiles.Enabled = false;
                 textBoxRawFiles.Enabled = false;
 
                 groupBoxQcOptions.Enabled = false;
+            }
+        }
+
+        private void checkBoxModeMs1_CheckedChanged(object sender, EventArgs e)
+        {
+            if (checkBoxModeMs1.Checked)
+            {
+                ckbxOutputChromatograms.Checked = true;
+                groupBoxChromatograms.Enabled = true;
+                checkBoxChroMs2.Enabled = false;
+            }
+            else
+            {
+                ckbxOutputChromatograms.Checked = false;
+                groupBoxChromatograms.Enabled = false;
+                checkBoxChroMs2.Enabled = false;
             }
         }
 
@@ -287,6 +305,22 @@ namespace RawToolsGUI
             {
                 checkBoxReporterFilterMGF.Enabled = checkBoxReporterIonFiltering.Checked;
                 groupBoxMgfOpts.Enabled = true;
+                ckbxOutputParse.Checked = true;
+            }
+            else
+            {
+                checkBoxReporterFilterMGF.Enabled = false;
+                groupBoxMgfOpts.Enabled = false;
+            }
+        }
+
+        private void ckbxOutputFaims_CheckedChanged(object sender, EventArgs e)
+        {
+            if (checkBoxFaims.Checked)
+            {
+                checkBoxReporterFilterMGF.Enabled = checkBoxReporterIonFiltering.Checked;
+                groupBoxMgfOpts.Enabled = true;
+                ckbxOutputParse.Checked = true;
             }
             else
             {
@@ -412,9 +446,9 @@ namespace RawToolsGUI
                 command = "RawTools.exe";
             }
                         
-            if (!checkBoxModeParse.Checked & !checkBoxModeQC.Checked)
+            if (!checkBoxModeParse.Checked & !checkBoxModeQC.Checked & !checkBoxModeMs1.Checked)
             {
-                MessageBox.Show("You need to select a Mode for RawTools. Please pick QC, Parse, or both.", "Info");
+                MessageBox.Show("You need to select a Mode for RawTools. Please pick QC, Parse, both, or MS1.", "Info");
                 return;
             }
 
@@ -442,11 +476,42 @@ namespace RawToolsGUI
                 return;
             }
 
+            if (checkBoxModeMs1.Checked)
+            {
+
+                if (!ckbxOutputChromatograms.Checked)
+                {
+                    MessageBox.Show("You haven't selected any chromatogram output. Please check chromatograms.");
+                    return;
+                }
+
+                arguments.Append(" -");
+
+                if (checkBoxModeMs1.Checked)
+                {
+                    arguments.Append("ms1");
+                }
+
+                if (ckbxOutputChromatograms.Checked)
+                {
+                    if (checkBoxChroMs1.Checked | checkBoxChroTIC.Checked | checkBoxChroBP.Checked)
+                    {
+                        arguments.Append(" -chro ");
+                    }
+
+                    if (checkBoxChroMs1.Checked) arguments.Append("1");
+                    if (checkBoxChroTIC.Checked) arguments.Append("T");
+                    if (checkBoxChroBP.Checked) arguments.Append("B");
+                }
+            }
+
+
+
             if (checkBoxModeParse.Checked)
             {
 
                 if (!ckbxOutputMGF.Checked & !ckbxOutputMetrics.Checked & !ckbxOutputChromatograms.Checked &
-                    !ckbxOutputParse.Checked & !ckbxOutputQuant.Checked & !checkBoxModeQC.Checked)
+                    !ckbxOutputParse.Checked & !ckbxOutputQuant.Checked & !checkBoxModeQC.Checked & !checkBoxFaims.Checked)
                 {
                     MessageBox.Show("You haven't selected any data output. Please choose something.");
                     return;
@@ -465,6 +530,11 @@ namespace RawToolsGUI
                 if (ckbxOutputMetrics.Checked)
                 {
                     arguments.Append("x");
+                }
+
+                if (checkBoxFaims.Checked)
+                {
+                    arguments.Append(" -faimsMgf");
                 }
 
                 if (checkBoxDataOutputDirectory.Checked)
@@ -578,6 +648,8 @@ namespace RawToolsGUI
             else
             {
                 groupBoxQuantOpt.Enabled = false;
+                ckbxOutputMGF.Checked = false;
+                checkBoxFaims.Checked = false;
             }
         }
 
@@ -724,6 +796,8 @@ namespace RawToolsGUI
 
             checkBoxModeParse.Checked = Pars.ParseMode;
 
+            checkBoxModeMs1.Checked = Pars.Ms1Mode;
+
             checkBoxModeQC.Checked = Pars.QcMode;
 
             checkBoxRefinePrecursor.Checked = Pars.RefinePrecursorMassCharge;
@@ -736,6 +810,7 @@ namespace RawToolsGUI
             ckbxOutputChromatograms.Checked = Pars.DataOutputChromatograms;
             ckbxOutputMetrics.Checked = Pars.DataOutputMetrics;
             ckbxOutputMGF.Checked = Pars.DataOutputMGF;
+            checkBoxFaims.Checked = Pars.FaimsOutputMgf;
             ckbxOutputParse.Checked = Pars.DataOutputParseMatrix;
             checkBoxDataOutputDirectory.Checked = Pars.DataOutputDirectory;
             
@@ -780,6 +855,8 @@ namespace RawToolsGUI
 
             Pars.QcMode = checkBoxModeQC.Checked;
 
+            Pars.Ms1Mode = checkBoxModeMs1.Checked;
+
             Pars.RefinePrecursorMassCharge = checkBoxRefinePrecursor.Checked;
             
             Pars.MinCharge = comboBoxMinCharge.Text;
@@ -788,6 +865,7 @@ namespace RawToolsGUI
             Pars.DataOutputChromatograms = ckbxOutputChromatograms.Checked;
             Pars.DataOutputMetrics = ckbxOutputMetrics.Checked;
             Pars.DataOutputMGF = ckbxOutputMGF.Checked;
+            Pars.FaimsOutputMgf = checkBoxFaims.Checked;
             Pars.DataOutputParseMatrix = ckbxOutputParse.Checked;
             Pars.DataOutputDirectory = checkBoxDataOutputDirectory.Checked;