+++ changelog.md
@@ -4,6 +4,17 @@ All notable changes to RawTools will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.5] 2022-04-25
+### Changes
+- Added the ability to process files with only MS1 scans (-ms1 flag).
+- Added the ability to create MGF output for individual CV values in FAIMS data (-faimsMgf).
+- Added a FAIMS CV column to the parse and quant matrix.
+- Added FAIMS values to the metrics output, if available.
+
+### Known bugs
+- When processing a non-FAIMS file, if the faimsMgf option is enabled, it will simply re-write the whole MGF.
+
+
 ## [2.0.4] 2022-02-01
 ### Changes
 - Updated ThermoRawFileReader to version 5.0.0.71.